Use standalone jobs for Wikibase Selenium tests
This started back in March 2021, the intent is to extract the Wikibase
Selenium tests to remove some minutes from the wmf-quibble-selenium
job.

Wikibase has a `selenium-test` which runs tests for "bridge" and "repo".
The former has been disabled by default back in August 2023 and whether
to re-enable it is T354841.

New job:
- wikibase-selenium

Bug: T287582
Change-Id: I5232e82e9ce0e23dc457d6072f1cd6bd79f69d49