Add and use job to deploy branch in node project
This is intended to allow pure vuejs projects to deploy to some
branch-hosting infrastructure, e.g. Netlify or Verecel.
I.e. it is intended to fullfill a similar purpose to
patchdemo.wmflabs.org, but for non-mediawiki js projects.

The specific implementation is left to the individual repository. By
doing that, we avoid strongly coupling WMF CI to a single provider like
Netlfiy and allow the individual repositories to choose a different one
if that works better for them (e.g. Vercel).

It should work together with I37a0b4a82b894671989105230b41ebdf650be3e5

Bug: T278706
Change-Id: I5c23b29111aeeaf5a43b872bf71a10e2a0de20da