python-jobs: Point sonarscanner at expected branch names
The existing configuration of the {name}-tox-docker-with-sonar job
had {branch} and $ZUUL_BRANCH set to the same value, which resulted
in sonar attempting to apply the coverage report for patch X to the
code for X-1, triggering errors due to the mismatch.

To resolve, mimic the sonarqube handling in job-templates.yaml used
for java projects in the python configuration. This removes duplicate
definitions of the sonar_args found in jobs extending
*-tox-docker-with-sonar and replaces them with a constant definition.

This is a followup to Ie1bb707f67 which appears to have introduced
this problem.

Bug: T283778
Change-Id: Ie07743cf2e98c6d7727359e2ce8f2c6db4c8d66b