dockerfiles: [node12-test] Do not process submodules
The image entry point erroneously process submodules using `--remote`,
that probably got carried from the -portals image which is an adhoc job.
That results in checking out the tracked repository tip of the branch
instead of the head registered in the submodule.

Moreover, the node12-test should not process any git checkout / fetch /
submodules, that is already handled by ci-src-setup.

Remove the git submodule command entirely.

Bug: T284492
Change-Id: I193ff70073fed608f6737761498799644f0b778b