tests: mute zuul output when testing layout
IntegrationTests.test_jjb_zuul_jobs runs zuul to validate the layout,
the server is set to use DEBUG level which would dump the whole layout
logic to the screen when we really only care about the failure message
dumped to stdout.

Just set logging level to WARNING which mutes Zuul. That results in way
less spam when the test fails.

To run the test:

Add a non existing job to a project in zuul/layout.yaml then:

  tox -e zuul_tests -- \
    tests/test_integration.py:IntegrationTests.test_jjb_zuul_jobs

Which list:

-------------------- >> begin captured stdout << ---------------------
FAILURE: Job job-that-does-not-exist-T232287 not defined

--------------------- >> end captured stdout << ----------------------

And a few lines of captured logging from jjb which is not the end of the
world.

Bug: T232287
Change-Id: Icbcd7868a63511cd7101ae3713279d0f80b2d56c