zuul: [mailman-templates] use i18n checker
The `debian-glue` job only runs for changes affecting a file in the
`./debian` directory, it is otherwise not triggered. Since L10n-bot only
touches `./i18n/*` the job is never run, the `gate-and-submit-l10n`
pipeline does not run any job and the change is left idling waiting for
job success that will never come.

Other repositories have the job `mediawiki-i18n-check-docker` which run
the diff through a large sed that attempts to detect HTML. Use it
instead of `debian-glue`.

Bug: T282022
Change-Id: Ie12fa4d4a3db67e94398b1442c4b0b6a11134d9c