wdqs-gui: Adjust build to expect configuration in docroot/
As part of generalizing wdqs into wikibase query service, to support
deployment of commons query service, expect configuration to be
maintained in the new sites/ directory. Maintains support for the
config file in the old location as well to ensure we don't have
broken builds during the transition. Once everything is merged
a followup can remove the exception for custom-config.json.

Bug: T280247
Change-Id: If993f188200ba8793fe40be9ca9c07f45cf94fce