jjb: Pass shellcheck at severity=critical
This portion of scripting is failing SC2157 in shellcheck: Argument to
implicit -n is always true due to literal strings. The script isn't
really wrong, it's just awkward due to having the external templating
write the shell script so shellcheck doesn't know it's a variable.

Make things explicit by assigning to a variable first, as a variable
shellcheck will know that this string can vary.

Change-Id: Ibbe05b8c994e716fc65cc4d6588f34fb33301b20