node14: Update from bundled npm 6 to pinned npm 7
I've updated the node10 file to keep it actually pinned at where it
was already at (it was missing a commit or tag specifier), and to
update the now "404 Not Found" /r/p/ URL.

I don't expect it to be rebuilt however. This is just to make it not
fail when locally building all images. No changelog increment.

Bug: T267888
Bug: T267889
Change-Id: Ib980f790410a618b9336edd7395db40745b6706d