zuul: polish set_doc_variables()
Rename raw_project to repo_name, when used set the destination to a
different variable "dest".

Set the default in an "else" clause instead of setting the default at
the top. Sounds clearer to fallback to the repository name after all
conditions have been exausted.

Add tests for mediawiki extensions and skins.

Change-Id: I74a4ff7dd85bd121a600a807afc2bf6b048218f3