utils/jjb-diff: support job filtering
When refactoring a set of job such as the Phan jobs, it is useful to be
able to compare just those jobs. That makes incremental edits way
faster.

In the jjb test command move the path argument further to the end and
pass any argument passed to jjb-diff. For example to compare solely the
phan jobs use:

 ./utils/jjb-diff '*phan*'

Takes 9 seconds versus 17 seconds for a full generation.

Change-Id: I81d76869ab73572dfe19929b2c724de9a8e6a4ad