jjb: play with Jinja2
Idea is to have the `docker-run` macro to accept a map of volumes to
mount and craft the extra command line arguments with a Jinja2 template.

This change is intended to NOT introduce any changes in the Jenkins
jobs, the templates thus have a few oddities related to whitespaces.
Notably, I found out that Jinja2 strips the trailing newline and
I thus wrapped the template in a dummy block which produces an extra
newline. I have send a patch to JJB upstream to flip the default:
https://review.opendev.org/c/jjb/jenkins-job-builder/+/818121

The other macros from jjb/macro-docker.yaml have been adjusted to no
more use `options: --volume=x:v` but instead the new system. A full
example would be:

 - builders:
    - docker-run:
        options: --user=nobody --entrypoint=/bin/ls
        args: -la /cache /srv/git
        volumes:
          cache:
          /srv/git: /srv/git:ro

Which would have the following bind mount from host to the container:

  | $WORKSPACE/cache | /cache
  | /srv/git         | /srv/git  (read-only)

A next change will migrate all occurences of `--volume` to volumes map.
We can later on do the same with `--env` or really any other option.

Change-Id: Ibb813b629e63addc0e8a04438d2536912e2b587c