jjb: normalize Quibble jobs to use /src
The intent is to remove the more specific JJB macro:
docker-run-with-log-and-workspace-cache and reduce the complexity.

Quibble runs in `/workspace` thus mount the agent `src` directory to
`/workspace/src`.  The macro now has a srcdir variable much like the
logdir one.

This cause the source and packages installed by Quibble to be moved on
the agents from `/var/lib/docker` to `/srv` partition.

Update quibble jobs.

Bug: T292729
Change-Id: I71a646a3172ab8f2849ced63419dc71bcdcad4d5