jjb: stop using host src for Quibble jobs
With ea5262de the Quibble jobs bind mounted $(pwd)/src inside the
container to prevent all the code to be inside the Docker container.
On the agents that is under /srv partition which is only 18G and when
big builds happen the partition fills up.

Since we now only have docker-run-with-log-cache-src macro which require
$(pwd)/src to be mounted, mount it to /unused. The Quibble jobs write to
/workspace/src which would be inside the container ad thus on the large
/var/lib/docker partition.

Rollback ea5262de

Bug: T292729
Change-Id: I1ac40a35b13a7d3304d602df5a9533fdf23a4371