Enforce editorconfig settings
Rename the `flake8` tox environment to `lint` and add
`editorconfig-checker` to run against all files in git.

Fix indentation of various files.

This is based on mediawiki/tools/train-dev change:
https://gerrit.wikimedia.org/r/c/736442

editorconfig-checker uses per lines regex and is unable to support
alignment space. Consider the following python code:

 print("foo"
       "bar")

The second line has six spaces which raises a complaint when the indent
size is set to 4. However the two spaces are purely for alignment and
are perfectly fine. Those alignments spaces are used in Debian
changelog, python and yaml files.

To support alignment space, run the checker twice:

1) without changelog, *.py, *.yaml files

2) against those files but ignoring the indentation check

For Python alignement is already covered by flake8. For Yaml if the
indentation is incorrect, it would cause an issue in our test suite and
will surely be caught.

':!*.py' is git pathspec described in gitglossary(7). All credits go to
Tyler Cipriani.

Change-Id: I701fd8a84ab32b0d2642d33e19a7ee40fef993f3