Ensure quibble image have the proper version
We sometime have the quibble images receiving a version bump which would
be ahead of Quibble. The latest example is 4807411a which added `npx`
command to the 1.4.0 Quibble images but bumped quibble-buster to 1.4.1.

Add a test which look at the version installed by pip in
releng/quibble-buster then lookup the Debian upstream version refered in
the changelog and ensure they are consistent.

The test can be run locally using:

  tox -e zuul_tests \
    tests/test_dockerfiles.py:test_quibble_images_version_is_in_sync

Change-Id: Iaaae8c340dddfd88616414bdb5356045871203b4