dockerfiles: renumber quibble-buster 1.4.1 to 1.4.0-s1
The Quibble Docker image version should be in sync with the Quibble
installed version. This never got enforced but will be with via test
added by Iaaae8c340dddfd88616414bdb5356045871203b4

4807411a added `npx` to the image and incorrectly bumped the
quibble-buster image from 1.4.0 to 1.4.1 instead of 1.4.0-s1. This apply
the proper version and the image has been manually retagged.

Change-Id: I8ddcdb448a78e3e93a37d70d59cb6284511f10fd