jjb: tie castor saving job to a label
To ease migration of the Castor service from an instance to another one,
bind the `castor-save-workspace-cache` job to a label rather than an
agent name. This way when a replacement agent is setup we can switch by
moving the label from an agent to another one instead of having to
update the job.

integration-agent-castor03 had the label applied:
https://integration.wikimedia.org/ci/computer/integration-castor03/

Update job castor-save-workspace-cache

Bug: T216244
Change-Id: Ia8f30e9a4353a6070480aa38617abbfd5d088e0d