dockerfiles: unhardcode MEDIAWIKI_URL and MEDIAWIKI_API_URL
The variables are used by the ruby selenium framework to reach the
MediaWiki web. They were hardcoded with port 9412 which is used for the
PHP built-in server, when switching to Apache on CI we have it listening
on port 9413.

Forge the environment variables using MW_SERVER and MW_SCRIPT_PATH which
are set by Quibble and come from the `--web-url` parameter.

Change-Id: Ifed15085282d56ba3b3a0718f7142d46aaa4d0de