jjb: Adjust phpbench args
The current setup is close, but not exactly what we need for aggregate
reports. This is a feature in phpbench that allows you to compare the
patch against the main branch, to check for performance regressions.

The --phpunit-aggregate flag will only get recognized if the quibble
stage is 'phpbench', but in the current config we are running a
direct command with "-c composer run -- phpbench".

Running `quibble --run=phpbench --phpbench-aggregate` does something
different (see the Phpbench class in commands.py); it will checkout
the previous commit, do a baseline check, then checkout HEAD
and run a comparison report.

So, this patch reworks the jjb config to run
`quibble --run=phpbench --phpbench-aggregate`.

Bug: T291549
Change-Id: Ifc6b8f682ab606853168807d91bdedf62e97faef