jjb: fix arguments passing in qemu-run.bash
To ssh/scp with the VM we have the options stored in a variable which I
double quoted with d99a82d492. That causes the whole list of options to
be passed as a single argument which causes:

  Bad port ' 4293 -i ./root.key -F /dev/null ...

Change the options to be stored in an array and expand it on usage.

Update jobs:
    fresh-test
    mw-cli-test

Change-Id: I64a0843dcfdac8a6df8ff1497f0a63351a29d504