dockerfiles: force color output for bazel
Bazel automatically detects whether colors should be output, most
probably based on whether stdin is /dev/null or is a tty. Jenkins
supports ANSI coloring output so we can turn it on.

In the global `bazelrc` set `--color=yes` for all commands which can be
done via the pseudo command `common`.

Change-Id: I8386929aa966adbab0558093109a89ca7ae74b2f