Do not test `labs/tools/heritage` with PHP 8
We started testing the PHP part of heritage in 8a66ecaa0
and bda0eb88bc, and moved to use Composer in c99668d5618.

This was upgraded to use `composer-php72-docker` in 1df31633,
and then at some point in time to `composer-test-php72-or-later`,
which included testing under PHP 8.

The Heritage repo does not support PHP 8, and we have no plans
of supporting it at this time. This version does not seem
available yet on Toolforge anyhow.

- Add job-template `composer-test-php72-or-later-but-not-php8`
- Use it instead of `composer-test-php72-or-later` for heritage

Bug: T316462
Change-Id: Id36cd8e5d5bbe51c007059612540a2627d5bc1a6