Add experimental composer apitests job to Wikibase
The Wikibase REST API team is waiting for security review on a third
party library (T316523), but would like to already start development of
the feature using this library. We rely heavily on the api-testing jobs
for e2e and schema tests, so it would be great to be able to have a
composer job that can run all tests for all features, including the one
depending on the library awaiting security review.

Bug: T316245
Change-Id: I34334161be1d1a78c04a24eb55c0c2185b8b728e