Always run Wikibase composer apitests in test step
Adding the job to experimental helped, and we found a (slightly hacky)
way to run certain API tests only if the required package is present
(Ide680f0800ad4acfdbc5615856decfd6d9946231). It looks like we'll still
have to wait a while for security review and we'd like to have it in the
test step of the pipeline in the meantime, so that we don't always have
to check experimental for related patches.

Change-Id: Ib645edfdf998c244823dfc3e3459d9a2260211b8