dockerfiles: Temporarily install buildctl from upstream master
We need buildctl to support the `registrytoken` auth config field for
performing JWT auth against our registry. This support is only present
in the upstream's master branch.

Note that a backport has been submitted upstream targeting the 0.10
branch.

https://github.com/moby/buildkit/pull/3150

If the PR is accepted and merged, we can revert this change and go back
to installation via the release tarball and validation via checksum.
Also note, however, that since Docker uses content addressable image
IDs, that approach is not all that different from copying directly from
of upstream images referenced by the image's SHA256, as Docker images
are content addressable. We could potentially continue use of `COPY
--from=<upstream image>` as the installation medium with the same degree
of assurance as performing checksum validation.

Bug: T308501
Change-Id: I6f016790cfe04b36f47bce6550ff657af2efcbe3