zuul: run PHPUnit standalone tests on REL branches
The PHPUnit Standalone group is not triggered on MediaWiki release
branches. When the `extension-quibble-only-standalone` job template got
introduced for T225068 we forgot to trigger the jobs in the REL specific
pipelines.

Add a test to ensure MediaWiki job templates (starts with 'extension-'
or 'skin-') have REL pipelines. It
shows several lack the release pipelines:

extension-apitests
extension-codehealth
extension-coverage
extension-gate
extension-javascript-documentation
extension-phpbench
extension-quibble-only-standalone
skin-coverage

Exclude all of the templates explicitly with the exception of
'extension-quibble-only-standalone' which is the topic of T322506. The
other can later on be added to release branches as needed
(extension-apitests sounds like a good candidate).

Affected repositories:

mediawiki/extensions/CirrusSearch
mediawiki/extensions/Scribunto
mediawiki/extensions/WikiLambda

Bug: T322506
Change-Id: I82b68dbaf67ef7588605161175a55ad66325d2ae