Zuul: [mediawiki/extensions/WikiLambda] Disable selenium tests
We're going to use e2e tests daily on betacluster instead.

For context, WikiLambda requires two services running to orchestrate and
evaluate functions. There currently is not a way to stand-up these
services in the quibble/jenkins/zuul CI system. (See
https://phabricator.wikimedia.org/T280070). Our team is working on a way
to run these tests in Helm, perhaps on the Gitlab pipeline. (See
https://phabricator.wikimedia.org/T321942).

This change turns off selenium testing as part of the pre-merge checks
as an interim solution.

Bug: T294388
Change-Id: Ic14f8cbf1623039917a0589f12d1fada4b2fa75b