zuul: Add words "sqlite" and "postgres" to "check php" description
I think it's good that we only publicly document one 'check' command
for a given pipeline to avoid creating a complicated impression,
however for discovery it is useful to mention that this covers additional
db types as well, not just alternative versions. Kind of thinking
about it from the perspective of PHPUnit coverage, not (only) PHP
versions.

Change-Id: I98fc9eebe2b7b43a05d6a187254877afa2653274