Add patch coverage job for parsoid
MediaWiki core, extensions and skins have jobs checking whether a change
enhances or degrage the test coverage. That is done running MediaWiki
phpunit.php wrapper against the proposed change and the parent patch and
report back the difference.

For MediaWiki, it is using Quibble, the phpunit.php wrapper plus some
editing of the phpunit XML configuration. The logic is in the image
releng/quibble-buster-php74-coverage

Create a job composer-coverage-patch-docker porting the logic from the
coverage image. It is using the base composer-php74 image which is good
enough for a standalone repository.

Add the job to `mediawiki/services/parsoid` in the `coverage` pipeline.

We can later on apply the job to all `mediawiki/libs/*` repositories by
adding it to the `mediawiki-php-library*` job templates. They already
are publishing coverage via `phpunit-coverage-php74-docker-publish`.

Change-Id: Ib40f7661fddafca357b782dd6e8cc3ee821e1b2c