jjb: Avoid "dubious ownership" git failure in mediawiki-i18n-check-docker
Git now checks to be sure the repo ownership matches that of the current
process. Since we're running mediawiki-i18n-check-docker in a docker
container as nobody, it should be fine to omit the current working
directory from this check by configuring `safe.directory`.

Avoid silent failure of checks by adding `set -e` and wrapping the
complex sed/grep calls in functions.

Bug: T329266
Change-Id: I82386c832646b6568bfb08cd30137daacbb130d2