jjb: use same order when creating directories
Ensure to have a consistent order creating the `log` directory first
followed by `src`. This will ease a future change which replaces common
code to use a single macro and will produce a smaller jjb diff.

We some time load the castor cache after creating the directories. Move
it to be the first action (in jjb/php.yaml).

Change-Id: I3d43b1d1dd0f4e1e056111022a218798d887b087