Remove gerrit-git-fat-pull job
I am migrating operations/software/gerrit deployment branch to rely on
`git-lfs` instead of `git-fat`. The reason for the pull job was that
sometime all artifacts were not passed to `deploy_artifacts.py` script
which resulted in missing artifacts when deploying. I have sometime
encountered an issue with Archiva not publishing the uploaded artifacts.

With `git-lfs` as long as the `.gitattributes` file is present which
kicks in LFS, one must have it installed and upon pushing for review the
artifacts get uploaded as well. There is thus no more need to verify
they are available.

Bug: T333465
Change-Id: Iee07603f3e5385505d3787c3970e7774e04cca86