dockerfiles: remove cassandra311
The `docker.wikimedia.org/releng/cassandra311` image was created as part
of T256281. A first attempt proposed it for SRE maintained images at
https://gerrit.wikimedia.org/r/c/operations/docker-images/production-images/+/607918/
but then moved to integration/config.

The image is based on Debian Stretch which is long EOL (T278203). From
discussion on tasks, the intent was to add some functional testing to
Kask but did not realize.

We can later recreate the image, though production-images repository
would probably be a better fit.

Bug: T335780
Bug: T224041
Change-Id: Ia479e63ee57117b2a5f86c9521e9df82498171a0