zuul: report using a different connection
To avoid a race condition between the source and reporter connections,
switch the reporters to the `gerrit-reporter` connection. The connection
must first be configured via Puppet.

Add a test to ensure we have no reporter using the same name as a
source.

The new connection should be available once Zuul scheduler is reloaded.

Zuul merger should not be affected, the items to merge gets the source
from `connection_name` which is feed by the Zuul scheduler from the
source connection name which is not changed.

Bug: T309376
Depends-On: I0e7369b8f6c533498c2abad75ab7da439a94572e
Change-Id: I92024031803b5aea861efcac6f9666cd46f80f36