Dockerfiles: [maven-java8] Update gerrit.wikimedia.org IP
When we recently migrated gerrit to new hardware it included a change
of the service IP behind gerrit.wikimedia.org.

The host key here has not changed since it's installed by puppet,

See "$sshkey = 'AAAAB3NzaC1yc2EAAAADAQABAAAAgQCF8p ..."
in BOTH zuul/merge.pp and manifests/gerrit.pp.
(ideally we wouldn't need this multiple times).

But the IP address changed for sure. The IP ending in .137
became gerrit-old.wikimedia.org and the IP ending in .151
is now gerrit.wikimedia.org.

For one reason or another this does not seem to have broken anything.

But let's fix it as part of removing the old server and IP that is bound
to its interface.

Bug: T326368
Bug: T336427
Change-Id: Id22f77e658b22d04c1e1e69111be848588dc401f