remove IPs from gerrit_ssh_host_key file
After a recent migration the service IP of
gerrit.wikimedia.org changed.

Instead of replacing it we are removing it
completely because we assume it is just meta data
and the actual key (which is puppetized) is the only
thing that matters.

And that recudes the number of places where we hardcode
IP addresses by one.

No changelog bump is done cause those are based on stretch
and are not buildable anymore.

Bug: T336427
Change-Id: I93d47c7e222e8ed401276f15ff88d12d6c184de0