Revert "Enable the parsoid extension when testing Flow"
This reverts commit 4bb5140cbcc55aa98b5b4c462f109c5b916f45db.

Reason for revert: As hashar said, this makes back-porting to
production any time a relevant extension does stuff that backwards-
incompatible (e.g. parser test changes) impossible. We ran into this
with TMH (a gated extension) today, and so I think the cost of this
isn't worth the benefit, sorry.

Change-Id: Ie299cd0bb92c24a1049c0b78ff86d8c1d46be819