dockerfiles: use extension base path for coverage report
The coverage report for extensions and skins is generated with a path
filter passed to PHPUnit. The path has a `tests/phpunit` suffix
introduced by a88c3e9a687f1 which relies on a convention.

For Wikibase, the tests are in the `client` and `repo` subdirectories
and are thus filtered out by PHPUnit.

Drop the `tests/phpunit` suffix to let PHPUnit discover any test.

Bug: T288396
Change-Id: Ib09e1e7cf02627e87f7b6e7c1524297a279b5ee6