Add experimental coverage job for Wikibase
The coverage image for MediaWiki extension expects tests to be in a
`test/phpunit` subdirectory which is not the case for Wikibase (since it
is really two extensions: `client` and `repo`). The coverage is not
generated (T288396)

Add a new job mwext-phpunit-coverage-docker using the new image which
only filter based on the extension name. It should thus discover
Wikibase tests.

Add the job as an experimental one for now.

Bug: T288396
Change-Id: I8a39d3f5f337e602ecc92425e3c0b4b8a35d0d3d