jjb: convert PCC job to a matrix job
The matrix job let us offer the user the combination of labels a job
will run on. It offers a way for end users to pick up different PCC
workers running different versions of Puppet.

We currently use the `puppet-compiler-node` label, I have made the job
to use version specific labels:
- puppet5-compiler-node
- puppet7-compiler-node

`puppet5-compiler-node` has been added to the existing Puppet 5
compilers.

The parent job is run on the contint* agents (gathered under the label
`productionAgents`).  Once the builds combinations have been
determinated, the builds run on a node having the selected label.

Since Jenkins can't convert from a Freestyle job to a Matrix job.
When updating the job, the old one MUST be deleted (which causes all
build history to be lost)

Bug: T236373
Change-Id: I881681de83ddf5cf3573fd88651e0cb91b18efac