wikidata-query-gui-build: run shell scripts in a container
this is done in order to avoid an issue similar to
the one observed in T328441.
The scripts were added to the QS GUI repository as part of If8869cb0f.
Once we've made sure that this works (and actually
resolvs the issue at hand), a similar patchset will be uploaded
for wikidata-query-builder-build.

Please see the comments for further notes and open questions.

Bug: T328543
Change-Id: Id272e206a4f59873c3f4a8906f8d29e072e5362b