Dockerfiles: add openssh-client to node test images
`wikidata/query/gui` has a Jenkins job to build the GUI artifacts and
upload them as a proposed change to Gerrit.

The build logic has been moved from the Jenkins job to a container image
by 4aa2002328 and while testing against the source change
https://gerrit.wikimedia.org/r/c/wikidata/query/gui/+/949833/ the build
fails:

  ssh -o StrictHostKeyChecking=no: 1: ssh: not found

This is similar to the build job for analytics refinery repo (T210271).
The git push over ssh requires the OpenSSH client to be present in the
image.

Add `openssh-client`to node test images and cascade.

Bug: T328543
Change-Id: I4821810ec9ffc74156662bd8b25e9d4fbfbd575f