jjb: wikidata-query-gui-build push to git from Jenkins
The commit of the build can not be pushed from within the container. The
process runs as `nobody` but the ssh agent is spawned by the Jenkins
agent and exposes the socket as `jenkins-deploy` (and expects the
connecting process to be run by that user).

This move back the git push back to the Jenkins job.

Bug: T328543
Change-Id: I6fa39c5baabbcc34b4389c446c6f9f33220488b4