jjb: remove zuul-cloner and submodules for mw core Phan job
Some branches of mediawiki/core can contain submodules which we do not
analyze as part of the Phan job.

Originally we disabled submodules processing by using `zuul-cloner`
(which does not processes submodules) but later on we added
`ci-src-setup-simple` which does!

Remove `zuul-cloner` and instruct `ci-src-setup-simple` to not process
submodules.

Update:
    mediawiki-core-php73-phan-docker
    mediawiki-core-php74-phan-docker
    mediawiki-core-php80-phan-docker
    mediawiki-core-php81-phan-docker

Bug: T44455
Bug: T278203
Change-Id: I287ae61c0d723dea85553ec72c528da6e63d30c9