dockerfiles: upgrade tox to 4.8.0
Tox v4 is a rewrite and although it remains mostly back compatible, the
devil is in the details.

There is an upgrade guide at https://tox.wiki/en/latest/upgrading.html
and I have listed a few issues on T345695.

`[tox:jenkins]` is no more supported. I have created a tox plugin that
would reinject the section settings (`integration/tox-jenkins-override`
repo). The plugin has to be installed along tox. Since it is not
released to Pypi, install from source and use the sha1 to pin the
version (T345607).

Install tox from source using 4.8.0 which is the last supporting
python3.7 from Buster.
Install virtualenv from source using 20.21.1 which is the last
supporting python 2.7.

Both versions are now set as `ARG` which makes it easier to find them. I
did the same change for the operations-puppet image via T345152.

Since I was confused as to whether tox v4 got already deployed, this
change the images version number to be the same as the tox version in
use (4.8.0). The sole exception is `zuul-cloner` which is not used for
running tox and does not depend on it.

This is a breaking change!

Bug: T345695
Bug: T345607
Depends-On: Ife15751da24f28a6479ec7bdd6a902c35f70ff63
Change-Id: Ief3c095f4a75592faa74a60508a1c17d8d95693b