Zuul: Enable 'tarball' extensions already in the main gate
- AbuseFilter
- Echo
- PageImages
- Thanks

These four extensions are already in the main 'gatedextensions' list
which is run for master and wmf branches, so these are already well-
tested and acceptable in terms of speed.

Practically, this change will be a no-op for all except 'release'
(REL1_XX) branches, which use the tarball list but not the main
'gate' list.

Change-Id: Iaaf7ecf0629bb6ff1d6cae8d8d3782c79a5700f9