dockerfiles: civicrm: install db with root auth
When creating the database with `mysqld_install_db`, MariaDB now
defaults the authentication method to use a socket which sets an invalid
root password (but still allow a process owned by root to connect).

Change it to the legacy authentication method (normal) which creates an
initial root user without password.

That broke the build when migrating from Stretch to Bullseye.

Bug: T307178
Change-Id: I1e2c5718055f25db7315a10a872c03f90849a377