archive-repo.py: add a venv for Levenshtein module
The utils/archive-repo.py relies on Levenshtein python module to be
installed, it is however not necessarily present on one machine.

We have a similar use case for running JJB or Zuul locally for which
virtualenv are defined and can be created by tox (see ./jenkins-jobs).

Add an `archive-repo` test environment to tox which installs the module
and execute the script passing it all arguments given. Example usage:

  tox -e archive-repo -- -h

If one run the script directly and the import fail, hint about using tox
to get the dependency and exit.

Note python-Levenshtein got renamed Levenshtein, I have adjusted the URL
in a comment of `utils/archive-repo.py`.

Change-Id: Ief2c4524cec5eb4a1ac94c26cdb18431626f294b