jjb: use a job template for wikidata-query-*-build
The wikidata-query-builder-build and wikidata-query-gui-build jobs share
almost the same logic, the only few things that changes are:

- ZUUL_PROJECT
- directory being wiped (src/build vs dist)
- npm entry point to build the project (grunt only_build vs build)
- build output directory (./src/build vs ./src/dist-deploy)

Refactor the two jobs to a job template which is invoked from a new jjb
project passing the different settings. That makes it easier to see the
differences.

While at it move the image name to a variable to avoid repeating it on
each command.

Affected jobs:
 - wikidata-query-builder-build
 - wikidata-query-gui-build

Change-Id: Ia88dcb37be03be21ef2b0f054e69fac73211d43d