dockerfiles: keep pip and wheels in Quibble images
We have end to end tests for `integration/quibble` which install Quibble
from source.  I am working on a change to get rid of `setup.py` in favor
of `pyproject.toml`: I68686e668c3bac7cd1cb16c14a31f415d705edf9.

That requires pip to be present in the Quibble images in order to
install Quibble in a CI job. python-pip and python-wheels were
previously purged to save up a bit of disk space in the image layer.

Change-Id: Ib19f2a8b6800bed0744ccab19d089e33baa6e03a