jjb: ls WORKSPACE after it got wiped
When a build is canceled, the next build would sometime fails to create
the cache/log directory:

  mkdir: cannot create directory ‘log’: Permission denied

I highly suspect it is left behind by the build/Docker and being owned
by root it prevents the next build from creating the directory, though
it is still deleted letting the next next build work.

I suspect it is a race condition with our macro deleting the file while
the container that uses log is still being deleted, and hence the log
dir is still owned by Docker which keeps it around.

Bug: T282893
Change-Id: Id173125c401b3a3f958f6b26730b9b3c55f01082