jjb: ls WORKSPACE before creating dirs
We create a combination of directories (cache, log, src) in the Jenkins
build workspace which are then mounted in the container. Sometime the
directory creation fails with a permission denied error.

I have suspected the previous build to have left behind some of the
directories, potentially owned by `root:root`. In 6143edb8 I have added
a postbuild script which does a `ls` after cleaning up the workspace. We
had a build failure today and the previous build on the same job had an
empty output, so the issue must be different.

Instead of having to track the previous build, add a `ls` before
creating the `cache`, `log` or `src` directories. Leave the post build
`ls` as a potential reference point.

Bug: T282893
Change-Id: I2859ee16ad50381f1986af1648959b46ef0ad9b0