jjb: rename puppet-compiler-node to puppet5-compiler-node
With the introduction of Puppet 7 catalogue compiler nodes we have
introduced a puppet7-compiler-node. To differentiate between Puppet 5
and 7 nodes I have added a puppet5-compiler-node label to the existing
workers in addition to puppet-compiler-node.

Change the label for the PCC job triggered by Zuul to the new label.
Once deployed `puppet-compiler-node` label can be removed from the PCC
workers.

Affected jobs:
 - operations-puppet-catalog-compiler-test

Bug: T236373
Change-Id: If10e420ac26fa6dbafd2f110cff7b0fae724158e