dockerfiles: disable git safe.directory
git aborts when the uid it is running as is different than the one
owning the git directory, that is to prevent privilege escalation by
executing git hook as a different user than the one owning the
repository.

On CI the workload is executed in a container which isolates the code or
any potential hooks. The workspace directory might be owned by a
different user causing git to fail:

- on CI that is created by the Jenkins agent user for historicaly reason

- Fresh would bind mount the source from a directory which is most
  probably owned by the user invoking the command rather than nobody.

Disable git `safe.directory` from the git system-wide configuration
/etc/gitconfig, the same way we enable protocol version 2.

Since this is in the base images, rebuild the whole fleet of images.

Bug: T354409
Change-Id: I08f61b35a9180bcf2bf495dd5b451ab275055347