dockerfiles: add git-lfs to base image
We have the use case to fetch material from git LFS and thus require
git-lfs to be available in ci-src-setup-simple. Rebuild that image after
adding the package to all the CI base images.

I am not bothering rebuilding EVERY images, since ci-src-setup-simple is
the sole one in which we need git LFS.

Change-Id: I8712c007e9bbfe8a27850bae233ce557903a1ffa