jjb: enhance mediawiki-i18n-check job
The job went to set `safe.directory $(pwd)` (T329266), that is
unfortunately resulting in a race condition trying to lock
`/mnt/home/jenkins-deploy/.gitconfig`.

The repository files are owned by `nobody` as created by
`ci-src-setup-simple`, thus move the `git show` command to run inside a
container as well. That let us remove the git config call that ran on
the host and resolve the race condition (T331098).

Remove the `mktemp` which leaves behind a temporary file on the Jenkins
agent hosts under `/tmp`. Instead write to a fixed file
`log/additions.txt` which resolves the leak. The file is guaranteed to
not exist via `docker-log-dir` which wipes the directory at build start.

Attach the `additions.txt` to the build. It is unlikely to serve any
purpose, but we expect files written under `log/` to be attached.

Add TODO entries to move the logic to a standalone container and
possibly use `grep` instead of `sed`.

Bug: T331098
Bug: T329266
Change-Id: I4b6a4df0d80a17f20708fc0e5a6c9e2221e5428d