Make Parsoid depend on ParserFunctions explicitly
Some tests in the core parser tests file that are marked as parsoid only
use the #if parser function and fail the integrated mode without the
extension installed.  These tests probably deserve to be rewritten with
a parser tests specific parser function, not moved to the extension's
repo.  For now, continue to depend on ParserFunctions.

Follows-Up: If0e2faad08074dbeb030392d09d60809b131f08d
Change-Id: I80745895f7e177909a339c208ae188a4eee37012