jjb: fix mediawiki-i18n-check-docker empty error
The check can fail without any output such as:

  + git show FETCH_HEAD -U0
  + grep '^+'
  Build step 'Execute shell' marked build as failure

And Raymond reported that is the case for the 2nd patch being sent by
the l10n-bot.  While testing the check with some old change, I had the
script erroring out with an empty output.  The reason is the change was
old and Zuul crafts a merge commit for it which thus result in an empty
output when using `git show`.

Based on our old git-changed-in-head script add a few options:

* --find-renames=100% renamed files that had a slight change would be
  considered modified and thus included
* -m show differences for merge commits ...
* first-parent ... but only follow the first parent commit

Bug: T331098
Change-Id: I0d20777049b4b5fb555963b4d3cf65c0b1192379