jjb: update-jars no more need to set the git remote
With Ieb6e460f3bc58d15e955b1aaf40b44bfbbfb9484, ci-src-setup now sets
the git remote before fetching and checking out. That let git LFS infer
the remote endpoint to fetch from and we no more need to set the remote
explicitly in the `analytics-refinery-update-jars-docker` job.

We previoudly had the job doing a git remote for git-fat which happens
later in the build. The migration to git LFS requires us to set the
remote earlier and since that is now done by the image, we can remove
the `git remote add` from the Jenkins job.

Bug: T328472
Change-Id: I70e311100de111446a1767d45a83a8f3d32a93f8